move metadata class from view model service
